requirements.txt
duckdb~=1.1.3
pyiceberg~=0.8.0
pyarrow~=18.1.0
apache-airflow-providers-amazon
apache-airflow~=2.10.2
pandas~=2.2.3
minio~=7.2.12
pyspark~=3.0.3
pyspark-stubs~=3.0.0.post3
pymongo~=4.10.1